revert some moving around
